dump working test impl

+++ ewc/vault-snapshot.sh
@@ -2,9 +2,10 @@
 
 # Variables
 VAULT_ADDR=${VAULT_ADDR}
-S3_BUCKET=${S3_BUCKET}
+S3_BUCKET_BASE_PATH=${S3_BUCKET_BASE_PATH}
 AWS_ACCESS_KEY_ID=${AWS_ACCESS_KEY_ID}
 AWS_SECRET_ACCESS_KEY=${AWS_SECRET_ACCESS_KEY}
+AWS_REGION=${AWS_REGION:-"eu-north-1"}
 
 # Check if VAULT_ADDR is set
 if [ -z "$VAULT_ADDR" ]; then
@@ -13,14 +14,8 @@ if [ -z "$VAULT_ADDR" ]; then
 fi
 
 # Check if VAULT_ADDR is set
-if [ -z "$VAULT_TOKEN" ]; then
-  echo "Error: VAULT_TOKEN is not set."
-  exit 1
-fi
-
-# Check if VAULT_ADDR is set
-if [ -z "$S3_BUCKET" ]; then
-  echo "Error: S3_BUCKET is not set."
+if [ -z "$S3_BUCKET_BASE_PATH" ]; then
+  echo "Error: S3_BUCKET_BASE_PATH is not set."
   exit 1
 fi
 
@@ -40,13 +35,24 @@ fi
 SA_TOKEN=$(cat /var/run/secrets/kubernetes.io/serviceaccount/token)
 
 # Authenticate with Vault using the Kubernetes auth method to obtain a Vault token
-VAULT_TOKEN=$(vault write -field=token auth/kubernetes/login \
+export VAULT_TOKEN=$(vault write -field=token auth/kubernetes/login \
   role=backup-cron-job \
   jwt=$SA_TOKEN)
 
 # Generate ISO 8601 compliant timestamp
-# (needed to use sed as couldn't make it work with '%:z' in date command)
-TIMESTAMP_ISO_8601=$(date +%Y-%m-%dT%H:%M:%S$(date +%z | sed 's/\(..\)$/:\1/'))
+# Check the current timezone offset
+TIMEZONE_OFFSET=$(date +%z)
+
+# Generate ISO 8601 compliant timestamp
+TIMESTAMP=$(date +%Y-%m-%dT%H:%M:%S)
+TIMEZONE_OFFSET=$(date +%z)
+
+if [ "$TIMEZONE_OFFSET" == "+0000" ]; then
+  TIMESTAMP_ISO_8601="${TIMESTAMP}Z"
+else
+  # (need to use sed as couldn't make it work with '%:z' in date command)
+  TIMESTAMP_ISO_8601="${TIMESTAMP}$(echo $TIMEZONE_OFFSET | sed 's/\(..\)$/:\1/')"
+fi
 
 SNAPSHOT_NAME="snapshot-$TIMESTAMP_ISO_8601.snap"
 
@@ -55,7 +61,12 @@ SNAPSHOT_NAME="snapshot-$TIMESTAMP_ISO_8601.snap"
 vault operator raft snapshot save /tmp/$SNAPSHOT_NAME
 
 # Upload to S3
-aws s3 cp /tmp/$SNAPSHOT_NAME s3://${S3_BUCKET}${SNAPSHOT_NAME}
+aws s3 cp /tmp/$SNAPSHOT_NAME s3://${S3_BUCKET_BASE_PATH}${SNAPSHOT_NAME} --region "${AWS_REGION}"
+
+if [ $? -ne 0 ]; then
+  echo "Error: Failed to upload snapshot to S3"
+  exit 1
+fi
 
 # Clean up
 rm /tmp/$SNAPSHOT_NAME
